requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
pydantic[email]==2.9.2
passlib[bcrypt]==1.7.4
pyjwt==2.9.0
python-dotenv==1.0.1
pytest==8.3.3
httpx==0.27.2
